Rename Config.output_dir and add test.db
