+++ .gitignore
@@ -160,7 +160,6 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
 .DS_Store
-*.db
 data/
 .*creds
 .netlify
