+++ tests/conftest.py
@@ -1,0 +1,10 @@
+import os
+
+from app.utils.config import Config
+from app.utils.constants import Constants
+
+
+def pytest_sessionstart(session):
+    # move tests/test.db to data/data.db
+    Config.db = os.path.join(Config.data, 'data.db')
+    os.rename(os.path.join(Constants.Paths.ROOT, 'tests', 'test.db'), Config.db)
